refactor(outfitter): slim upgrade orchestration boundaries (OS-434)
